requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
ez-setup==0.9
gunicorn==20.1.0
python-environ==0.4.54
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
